Publicly expose TypedTransactions
