+++ eth_account/account.py
@@ -56,9 +56,6 @@ from eth_account._utils.signing import (
     to_standard_signature_bytes,
     to_standard_v,
 )
-from eth_account._utils.typed_transactions import (
-    TypedTransaction,
-)
 from eth_account.datastructures import (
     SignedMessage,
     SignedTransaction,
@@ -76,6 +73,9 @@ from eth_account.messages import (
 )
 from eth_account.signers.local import (
     LocalAccount,
+)
+from eth_account.typed_transactions import (
+    TypedTransaction,
 )
 
 VRS = TypeVar("VRS", bytes, HexStr, int)
