+++ eth_account/typed_transactions/blob_transactions/blob_transaction.py
@@ -40,20 +40,20 @@ from eth_account._utils.transaction_utils import (
     transaction_rlp_to_rpc_structure,
     transaction_rpc_to_rlp_structure,
 )
-from eth_account._utils.typed_transactions.access_list_transaction import (
-    access_list_sede_type,
-)
-from eth_account._utils.typed_transactions.base import (
-    TYPED_TRANSACTION_FORMATTERS,
-    Blob,
-    BlobPooledTransactionData,
-    _TypedTransactionImplementation,
-)
 from eth_account._utils.validation import (
     LEGACY_TRANSACTION_VALID_VALUES,
     is_int_or_prefixed_hexstr,
     is_rpc_structured_access_list,
     is_sequence_of_bytes_or_hexstr,
+)
+from eth_account.typed_transactions.access_list_transaction import (
+    access_list_sede_type,
+)
+from eth_account.typed_transactions.base import (
+    TYPED_TRANSACTION_FORMATTERS,
+    Blob,
+    BlobPooledTransactionData,
+    _TypedTransactionImplementation,
 )
